+++ scripts/aws/run_aws.sh
@@ -2,7 +2,14 @@
 # scripts/aws/run_aws.sh
 # 2023-02-02 | CR
 #
-set -o allexport ; . .env ; set +o allexport ;
+
+REPO_BASEDIR="`pwd`"
+cd "`dirname "$0"`"
+SCRIPTS_DIR="`pwd`"
+cd "${REPO_BASEDIR}"
+
+# set -o allexport ; . .env ; set +o allexport ;
+. ${SCRIPTS_DIR}/../set_app_dir_and_main_file.sh
 
 if [ "${APP_NAME}" = "" ]; then
     echo "APP_NAME not set"
@@ -14,39 +21,9 @@ if [ "${CURRENT_FRAMEWORK}" = "" ]; then
     exit 1
 fi
 
-export APP_NAME_LOWERCASE=$(echo ${APP_NAME} | tr '[:upper:]' '[:lower:]')
-
-# Default App main code directory
-if [ "${APP_DIR}" = "" ]; then
-  # https://aws.github.io/chalice/topics/packaging.html
-  APP_DIR='.'
-  if [ "${CURRENT_FRAMEWORK}" = "fastapi" ]; then
-    # https://fastapi.tiangolo.com/tutorial/bigger-applications/?h=directory+structure#an-example-file-structure
-    APP_DIR='app'
-  fi
-  if [ "${CURRENT_FRAMEWORK}" = "flask" ]; then
-    # https://flask.palletsprojects.com/en/2.3.x/tutorial/layout/
-    APP_DIR='flaskr'
-  fi
-fi
-
 if [ ! -d "./${APP_DIR}" ]; then
   echo "ERROR: APP_DIR './${APP_DIR}' not found"
   exit 1
-fi
-
-# Default App entry point code file
-if [ "${APP_MAIN_FILE}" = "" ]; then
-  # https://aws.github.io/chalice/topics/packaging.html
-  APP_MAIN_FILE='app'
-  if [ "${CURRENT_FRAMEWORK}" = "fastapi" ]; then
-    # https://fastapi.tiangolo.com/tutorial/bigger-applications/?h=directory+structure#an-example-file-structure
-    APP_MAIN_FILE='main'
-  fi
-  if [ "${CURRENT_FRAMEWORK}" = "flask" ]; then
-    # https://flask.palletsprojects.com/en/2.3.x/tutorial/factory/
-    APP_MAIN_FILE='__init__'
-  fi
 fi
 
 if [ ! -f "${APP_DIR}/${APP_MAIN_FILE}.py" ]; then
@@ -54,6 +31,7 @@ if [ ! -f "${APP_DIR}/${APP_MAIN_FILE}.py" ]; then
   exit 1
 fi
 
+export APP_NAME_LOWERCASE=$(echo ${APP_NAME} | tr '[:upper:]' '[:lower:]')
 
 AWS_STACK_NAME='${APP_NAME_LOWERCASE}-be-stack'
 
@@ -65,11 +43,6 @@ SSL_CA_CERT_PATH="./ca.crt"
 # RUN_METHOD="gunicorn"
 # RUN_METHOD="chalice"
 RUN_METHOD="chalice_docker"
-
-REPO_BASEDIR="`pwd`"
-cd "`dirname "$0"`"
-SCRIPTS_DIR="`pwd`"
-cd "${REPO_BASEDIR}"
 
 echo "SCRIPTS_DIR: ${SCRIPTS_DIR}"
 echo "REPO_BASEDIR: ${REPO_BASEDIR}"
@@ -143,9 +116,10 @@ if [[ "$1" = "run_local" || "$1" = "" ]]; then
     cd ${REPO_BASEDIR}
 
     echo ""
-    echo "Stage = ${STAGE}"
-    echo "Port = ${PORT}"
-    echo "RUN_METHOD = ${RUN_METHOD}"
+    echo "Stage: ${STAGE}"
+    echo "Port: ${PORT}"
+    echo "Run method (RUN_METHOD): ${RUN_METHOD}"
+    echo "Python entry point (APP_DIR.APP_MAIN_FILE): ${APP_DIR}.${APP_MAIN_FILE}"
     echo ""
 
     export IP_ADDRESS=$(sh ${SCRIPTS_DIR}/../get_localhost_ip.sh)
@@ -158,28 +132,28 @@ if [[ "$1" = "run_local" || "$1" = "" ]]; then
 
     echo "Run over: 1) http, 2) https, 3) use current [${RUN_METHOD}] (1/2/3) ?"
     read RUN_PROTOCOL
-    while [[ ! $RUN_PROTOCOL =~ ^[123]$ ]]; do
+    while [[ ! ${RUN_PROTOCOL} =~ ^[123]$ ]]; do
         echo "Please enter 1 or 2"
         read RUN_PROTOCOL
     done
-    if [ $RUN_PROTOCOL = "1" ]; then
+    if [ "${RUN_PROTOCOL}" = "1" ]; then
         RUN_PROTOCOL="http"
     else
         RUN_PROTOCOL="https"
     fi
 
     if [ "${CURRENT_FRAMEWORK}" = "chalice" ]; then
-        if [ $RUN_PROTOCOL = "https" ]; then
-            RUN_METHOD="chalice_docker"
+        if [ ${RUN_PROTOCOL} = "https" ]; then
+            export RUN_METHOD="chalice_docker"
         else
-            RUN_METHOD="chalice"
+            export RUN_METHOD="chalice"
             make down_qa
         fi
     else
-        if [ $RUN_PROTOCOL = "http" ]; then
+        if [ ${RUN_PROTOCOL} = "http" ]; then
             make down_qa
             echo "NOTE: The warning '-i used with no filenames on the command line, reading from STDIN.' is normal..."
-            APP_CORS_ORIGIN=$(echo ${APP_CORS_ORIGIN} | perl -i -pe 's|https:\/\/|http:\/\/|')
+            export APP_CORS_ORIGIN=$(echo ${APP_CORS_ORIGIN} | perl -i -pe 's|https:\/\/|http:\/\/|')
         fi
     fi
 
@@ -198,23 +172,9 @@ if [[ "$1" = "run_local" || "$1" = "" ]]; then
     fi
 
     if [ "${RUN_METHOD}" = "gunicorn" ]; then
-        if [ $RUN_PROTOCOL = "https" ]; then
-            echo "pipenv run run gunicorn --bind 0.0.0.0:${PORT} ${APP_DIR}.${APP_MAIN_FILE}:app --reload --certfile=${SSL_CERT_PATH} --keyfile=${SSL_KEY_PATH} --forwarded-allow-ips=${IP_ADDRESS}"
-            echo ""
-            pipenv run gunicorn ${APP_DIR}.${APP_MAIN_FILE}:app \
-                --bind 0.0.0.0:${PORT} \
-                --reload \
-                --workers=2 \
-                --certfile="${SSL_CERT_PATH}" \
-                --keyfile="${SSL_KEY_PATH}" \
-                --ciphers="TLSv1.2" \
-                --proxy-protocol \
-                --limit-request-field_size=200000 \
-                --forwarded-allow-ips="${IP_ADDRESS},127.0.0.1,0.0.0.0" \
-                --do-handshake-on-connect \
-                --strip-header-spaces \
-                --env PORT=${PORT} \
-                --env APP_STAGE="${STAGE}"
+        if [ ${RUN_PROTOCOL} = "https" ]; then
+            echo "${SCRIPTS_DIR}/../secure_local_server/run.sh"
+            ${SCRIPTS_DIR}/../secure_local_server/run.sh "run" ${STAGE}
         else
             echo "pipenv run run gunicorn --bind 0.0.0.0:${PORT} ${APP_DIR}.${APP_MAIN_FILE}:app --reload  --forwarded-allow-ips=${IP_ADDRESS}"
             echo ""
@@ -233,10 +193,9 @@ if [[ "$1" = "run_local" || "$1" = "" ]]; then
     fi
 
     if [ "${RUN_METHOD}" = "uvicorn" ]; then
-        if [ $RUN_PROTOCOL = "https" ]; then
-            echo "pipenv run uvicorn ${APP_DIR}.${APP_MAIN_FILE}:app --ssl-keyfile=${SSL_KEY_PATH} --ssl-certfile=${SSL_CERT_PATH} --reload --host 0.0.0.0 --port ${PORT}"
-            echo ""
-            pipenv run uvicorn ${APP_DIR}.${APP_MAIN_FILE}:app --ssl-keyfile=${SSL_KEY_PATH} --ssl-certfile=${SSL_CERT_PATH} --reload --host 0.0.0.0 --port ${PORT}  # --ca-certs=${SSL_CA_CERT_PATH}
+        if [ ${RUN_PROTOCOL} = "https" ]; then
+            echo "${SCRIPTS_DIR}/../secure_local_server/run.sh"
+            ${SCRIPTS_DIR}/../secure_local_server/run.sh "run" ${STAGE}
         else
             echo "pipenv run uvicorn ${APP_DIR}.${APP_MAIN_FILE}:app  --reload --host 0.0.0.0 --port ${PORT}"
             echo ""