+++ scripts/dns/run_local_dns.sh
@@ -16,6 +16,9 @@ if [ "${APP_NAME}" = "" ]; then
     echo "APP_NAME environment variable must be defined"
     exit 1
 fi
+
+export TMP_WORKING_DIR="/tmp"
+
 export APP_NAME_LOWERCASE=$(echo ${APP_NAME} | tr '[:upper:]' '[:lower:]')
 
 # These settings can be overwritten by the ".env" file
@@ -99,9 +102,9 @@ fi
 if [ "${ACTION}" = "" ]; then
 
     # This will be added to the DNS configuration file "/etc/bind/named.conf"
-    echo "Creating ${SCRIPTS_DIR}/dns/config/named-to-add.conf"
-    mkdir -p ${SCRIPTS_DIR}/dns/config
-    cat > ${SCRIPTS_DIR}/dns/config/named-to-add.conf <<EOF
+    echo "Creating ${TMP_WORKING_DIR}/dns/config/named-to-add.conf"
+    mkdir -p ${TMP_WORKING_DIR}/dns/config
+    cat > ${TMP_WORKING_DIR}/dns/config/named-to-add.conf <<EOF
 zone "${DNS_DOMAIN_NAME}" {
     type master;
     file "/etc/bind/zones/${DNS_DOMAIN_NAME}";
@@ -109,9 +112,9 @@ zone "${DNS_DOMAIN_NAME}" {
 EOF
 
     # This will be the domain zone configuration file
-    echo "Creating ${SCRIPTS_DIR}/dns/config/zones/${DNS_DOMAIN_NAME}"
-    mkdir -p ${SCRIPTS_DIR}/dns/config/zones
-    cat > ${SCRIPTS_DIR}/dns/config/zones/${DNS_DOMAIN_NAME} <<EOF
+    echo "Creating ${TMP_WORKING_DIR}/dns/config/zones/${DNS_DOMAIN_NAME}"
+    mkdir -p ${TMP_WORKING_DIR}/dns/config/zones
+    cat > ${TMP_WORKING_DIR}/dns/config/zones/${DNS_DOMAIN_NAME} <<EOF
 \$TTL    604800
 @       IN      SOA     ${DNS_DOMAIN_NAME}. admin.${DNS_DOMAIN_NAME}. (
                             2023112601 ; Serial
@@ -123,9 +126,11 @@ EOF
 @       IN      NS      ${DNS_DOMAIN_NAME}.
 @       IN      A       ${IP_ADDRESS}
 EOF
-    echo "Creating ${SCRIPTS_DIR}/dns/Dockerfile"
-    cp ${SCRIPTS_DIR}/dns/Dockerfile.template ${SCRIPTS_DIR}/dns/Dockerfile
-    perl -i -pe "s|APP_NAME_LOWERCASE_placeholder|${APP_NAME_LOWERCASE}|g" "${SCRIPTS_DIR}/dns/Dockerfile"
+    echo "Copying ${TMP_WORKING_DIR}/dns/docker-compose.yml"
+    cp ${SCRIPTS_DIR}/dns/docker-compose.yml ${TMP_WORKING_DIR}/dns/docker-compose.yml
+    echo "Creating ${TMP_WORKING_DIR}/dns/Dockerfile"
+    cp ${SCRIPTS_DIR}/dns/Dockerfile.template ${TMP_WORKING_DIR}/dns/Dockerfile
+    perl -i -pe "s|APP_NAME_LOWERCASE_placeholder|${APP_NAME_LOWERCASE}|g" "${TMP_WORKING_DIR}/dns/Dockerfile"
 
     if ! docker ps > /dev/null 2>&1;
     then
@@ -157,7 +162,7 @@ EOF
     then
         docker restart dns-server
     else
-        docker-compose -f ${SCRIPTS_DIR}/dns/docker-compose.yml up -d
+        docker-compose -f ${TMP_WORKING_DIR}/dns/docker-compose.yml up -d
     fi
 
     # Refresh the forntend/backend ".env" files to reflect the new domain name